+++ src/main/java/no/taardal/blossom/entity/Entity.java
@@ -5,8 +5,8 @@ import no.taardal.blossom.keyboard.Keyboard;
 
 public abstract class Entity {
 
-    protected double x;
-    protected double y;
+    protected int x;
+    protected int y;
     protected double velocityX;
     protected double velocityY;
 
@@ -14,19 +14,19 @@ public abstract class Entity {
 
     public abstract void draw(Camera camera);
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
